config modal system

+++ src/components/MacroItem/index.jsx
@@ -3,7 +3,7 @@ import Container, { Delete, Edit, Name, Command } from './styled';
 import { Icon } from '@fluentui/react/lib/Icon';
 import { Text } from '@fluentui/react';
 
-const MacroItem = ({ _id, actived, name }) => {
+const MacroItem = ({ _id, actived, name, setModalConfig }) => {
 	const switchActived = () => {
 		fetch(`http://localhost:8080/macro/switchActived/?id=${_id}`, {
 			method: 'PUT',
@@ -19,19 +19,28 @@ const MacroItem = ({ _id, actived, name }) => {
 			});
 	};
 
-  const deleteMacro = () => {
-		fetch(`http://localhost:8080/macro/delete/?id=${_id}`, {
-			method: 'DELETE',
-			body: JSON.stringify({ actived: !actived }),
-			mode: 'cors',
-			headers: {
-				'Content-Type': 'application/json',
+	const deleteMacro = () => {
+		setModalConfig((prevState) => ({
+			...prevState,
+			title: `Delete ${name}`,
+			subText: 'Are you sure you want to delete this macro?',
+			show: true,
+			confirmFunction: () => {
+				fetch(`http://localhost:8080/macro/delete/?id=${_id}`, {
+					method: 'DELETE',
+					body: JSON.stringify({ actived: !actived }),
+					mode: 'cors',
+					headers: {
+						'Content-Type': 'application/json',
+					},
+				})
+					.then((response) => response.json())
+					.then(({ data }) => {
+						console.log(data);
+					});
+				setModalConfig((prevState) => ({ ...prevState, show: false }));
 			},
-		})
-			.then((response) => response.json())
-			.then(({ data }) => {
-				console.log(data);
-			});
+		}));
 	};
 
 	return (
